requirements.txt
# Dependencias del Procesador de Archivos BLF
# Instalar con: pip install -r requirements.txt

# Procesamiento de archivos CAN
cantools>=40.5.0
python-can>=4.6.0

# Manipulación de datos
pandas>=2.0.0
numpy>=1.20.0

# Visualización
matplotlib>=3.5.0
pyqtgraph>=0.13.0

# Interfaz gráfica
PyQt5>=5.15.0

# Utilidades adicionales
argparse-addons>=0.12.0
bitstruct>=8.20.0